requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
Gunicorn==20.0.4
Numpy==1.19.1
Pillow==7.2.0